# Stash methods that get and set thread local variables to allow the user to mock them