Fixes #605.
Alternative to #581 to fix #580