@pirj 2024-07-13
@ioquatix we fixed a side issue of not being able to mock get_instance_variable. Does it still work for your uses?