@JonRowe 2024-07-13
I would argue that you shouldn't really be mocking the internals of thread, its not something you own, and this implementation still wouldn't allow you to mock `Thread.current` which is more likely IMO that individual methods...