@ioquatix 2024-07-13
@pirj I stopped using RSpec so I don't have any strong opinion about it. However, I still think it would be better to use `Thread.attr_accessor :rspec_state` or something like that, would have also avoided this issue...