@manueljacob 2024-07-13
> Can you please add a spec that would fail before your changes?

Done. Not sure how I could miss this.