@manueljacob 2024-07-15
> LGTM, however have you measured the performance overhead?

I pushed some performance optimization (using `UnboundMethod#bind_call`) instead of separate `UnboundMethod#bind` and `Method#call`, and did some quick tests. I replaced stdout and stderr to avoid writing 1000000 dots to the terminal.

```ruby
require 'benchmark'
require 'rspec'

out = StringIO.new

puts Benchmark::CAPTION
puts(Benchmark.measure do
  RSpec.describe do
    1000000.times do
      it do
      end
    end
  end

  RSpec::Core::Runner.run([], out, out)
end)

out.seek(0)
puts out.readlines[-3..-2]
```

Before this PR:
```
      user     system      total        real
 85.016527   2.258209  87.274736 ( 87.424472)
Finished in 59.82 seconds (files took 27.66 seconds to load)
1000000 examples, 0 failures
ruby rspec_perf.rb  86,45s user 2,69s system 99% cpu 1:29,46 total
```

After this PR:
```
      user     system      total        real
 85.154865   2.142675  87.297540 ( 87.447185)
Finished in 59.88 seconds (files took 27.65 seconds to load)
1000000 examples, 0 failures
ruby rspec_perf.rb  86,53s user 2,60s system 99% cpu 1:29,47 total
```

From these results, it’s only very slightly slower. However, the test conditions were not perfect (light background activity, power saving enabled).