@manueljacob 2024-07-15
`UnboundMethod#bind_call` was added in Ruby 2.7, so I re-added the `UnboundMethod#bind` / `Method#call` code for older Ruby versions. On Ruby 2.6, I observed a much larger slowdown of up to 10% on the above benchmark. However, variance is high on my current benchmarking setup.

To do some proper benchmarking, I should follow [that](https://github.com/google/benchmark/blob/main/docs/reducing_variance.md) and [that](https://llvm.org/docs/Benchmarking.html). However, currently I don’t have a suitable machine available for that.