@ioquatix 2024-07-15
Nice work, I assume pre-bind_call is more than trivially slower?